+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2020-01-06
+
+- **Added** pages from legacy WordPress blog:
+    - [intellij-idea-cheat-sheet](https://codyburleson.com/intellij-idea-cheat-sheet)
+- **Fixes** [#1](https://github.com/codyburleson/cburleson-pwa/issues/1)Unable to select and copy text on desktop
+
 ## [1.2.8] - 2020-01-04
 
 - **Added** pages from legacy WordPress blog: